pyproject.toml
[tool.poetry]
name = "polyvoxcord"
version = "0.1.0"
description = ""
authors = ["somakai_sumasi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = {extras = ["voice"], version = "^2.3.1"}
python-dotenv = "^1.0.0"
pandas = "^2.0.3"
sqlalchemy = "^2.0.20"
requests = "^2.31.0"
pyvcroid2 = {git = "https://github.com/Nkyoku/pyvcroid2.git"}
isort = "^5.12.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
